Fixed the memory leak associated with not stopping the `eth_getFilterChanges` requests (#232)
* refactoring the Scanner class

* refactoring Scanner

* pre-merge

* inheritance

* rm watcher.ts

* fixed

* fix

* fix
